# openapi.yaml
openapi: 3.0.3
info:
  title: WebAppX API
  description: Complete API for WebAppX marketplace platform
  version: 1.0.0
  contact:
    name: API Support
    email: support@webappx.com
servers:
  - url: http://localhost:5173/webappx
    description: Development server
  - url: https://api.webappx.com
    description: Production server
paths:
...
  /api/login:
    post:
      summary: Authenticate user
      description: Authenticate a user with username and password, returning a token and user information. This endpoint is publicly accessible and does not require authorization.
      tags:
        - Authentication
      security: []  # No security required - public endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              admin_user:
                summary: Admin user login
                value:
                  username: "alice"
                  password: "alice"
              buyer_user:
                summary: Buyer user login
                value:
                  username: "bob"
                  password: "bob"
              seller_user:
                summary: Seller user login
                value:
                  username: "carol"
                  password: "carol"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/me:
    get:
      summary: Get current user profile
      description: Retrieve the current user's profile information using the authentication token
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # === SETTINGS ===
  /api/settings:
    get:
      summary: Get application settings
      description: Retrieve all application settings
      tags:
        - Settings
      security: []  # Public endpoint
      responses:
        '200':
          description: Settings retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Setting'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # === ADMIN DASHBOARD ===
  /api/admin/dashboard:
    get:
      summary: Get admin dashboard statistics
      description: Retrieve dashboard statistics for admin users
      tags:
        - Admin
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Dashboard statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # === ADMIN USERS ===
  /api/admin/users:
    get:
      summary: Get all users
      description: Retrieve all users (admin only)
      tags:
        - Admin
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/admin/users/{id}:
    patch:
      summary: Update user status
      description: Update user status (ban/unban) - admin only
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
              properties:
                action:
                  type: string
                  enum: ["toggleBan"]
                  description: Action to perform on the user
            example:
              action: "toggleBan"
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # === ADMIN SELLERS ===
  /api/admin/sellers:
    get:
      summary: Get all sellers
      description: Retrieve all sellers (admin only)
      tags:
        - Admin
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sellers retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Seller'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/admin/sellers/{id}:
    patch:
      summary: Update seller status
      description: Approve, reject, activate, or deactivate a seller - admin only
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Seller ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
              properties:
                action:
                  type: string
                  enum: ["approve", "reject", "activate", "deactivate"]
                  description: Action to perform on the seller
            examples:
              approve:
                summary: Approve seller
                value:
                  action: "approve"
              reject:
                summary: Reject seller
                value:
                  action: "reject"
      responses:
        '200':
          description: Seller updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Seller'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # === ADMIN PRODUCTS ===
  /api/admin/products:
    get:
      summary: Get all products
      description: Retrieve all products (admin only)
      tags:
        - Admin
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/admin/products/{id}:
    patch:
      summary: Update product status
      description: Approve, reject, flag, or remove a product - admin only
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Product ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
              properties:
                action:
                  type: string
                  enum: ["approve", "reject", "flag", "remove"]
                  description: Action to perform on the product
            examples:
              approve:
                summary: Approve product
                value:
                  action: "approve"
              remove:
                summary: Remove product
                value:
                  action: "remove"
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # === ADMIN REPORTS ===
  /api/admin/reports:
    get:
      summary: Get all reports
      description: Retrieve all reports (admin only)
      tags:
        - Admin
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Reports retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Report'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/admin/reports/{id}:
    patch:
      summary: Resolve report
      description: Mark a report as resolved - admin only
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Report ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
              properties:
                action:
                  type: string
                  enum: ["resolve"]
                  description: Action to perform on the report
            example:
              action: "resolve"
      responses:
        '200':
          description: Report resolved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # === ADMIN SETTINGS ===
  /api/admin/settings:
    get:
      summary: Get admin settings
      description: Retrieve admin settings (admin only)
      tags:
        - Admin
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Settings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                example:
                  siteName: "MyMarketplace"
                  currency: "USD"
                  payoutDelayDays: "7"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update admin settings
      description: Update admin settings (admin only)
      tags:
        - Admin
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
            example:
              siteName: "MyMarketplace"
              currency: "USD"
              payoutDelayDays: "7"
      responses:
        '200':
          description: Settings updated successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # === SELLER PRODUCTS ===
  # Note: Seller products are the same as regular products, but with seller-specific access controls.
  # Only the seller who created the product can view, update, or delete their own products.
...
components:
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: User's username
          example: "alice"
          minLength: 1
          maxLength: 50
        password:
          type: string
          description: User's password
          example: "alice"
          minLength: 1
          maxLength: 100
    LoginResponse:
      type: object
      required:
        - token
        - user
      properties:
        token:
          type: string
          description: JWT token for authentication
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: '#/components/schemas/PublicUser'

    # === USER MODELS ===
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing_credentials:
              summary: Missing credentials
              value:
                MESSAGE: "Username and password are required"
            invalid_format:
              summary: Invalid format
              value:
                MESSAGE: "Invalid username or password format"
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_credentials:
              summary: Invalid credentials
              value:
                MESSAGE: "Invalid username or password"
            invalid_token:
              summary: Invalid token
              value:
                MESSAGE: "Invalid authentication token"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            server_error:
              summary: Server error
              value:
                MESSAGE: "Internal server error"
            database_error:
              summary: Database error
              value:
                MESSAGE: "Database connection failed"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
...